@@ -4,8 +4,11 @@
   "private": true,
   "scripts": {
     "docusaurus": "docusaurus",
-    "start": "docusaurus start  --host 0.0.0.0",
-    "build": "docusaurus build",
+    "start": "docusaurus start",
+    "build": "export NODE_OPTIONS=--max-old-space-size=16000 && docusaurus build",
+    "dev": "set NODE_OPTIONS=--max-old-space-size=16000 && docusaurus start --host 0.0.0.0 --no-open",
+    "build:win": "set NODE_OPTIONS=--max-old-space-size=16000 && docusaurus build --config dev.config.js",
+    "check": "npx docusaurus-mdx-checker",
     "swizzle": "docusaurus swizzle",
     "deploy": "docusaurus deploy",
     "clear": "docusaurus clear",
@@ -14,17 +17,19 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.3",
-    "@docusaurus/preset-classic": "^2.4.3",
-    "@mdx-js/react": "^1.6.22",
+    "@docusaurus/core": "^3.6.3",
+    "@docusaurus/preset-classic": "^3.6.3",
+    "@gera2ld/jsx-dom": "^2.2.2",
+    "@mdx-js/react": "^3.0.0",
     "clsx": "^1.2.1",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-player": "2.10.1"
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "use-local-storage-state": "^19.4.1"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.3"
+    "@docusaurus/module-type-aliases": "^3.6.3",
+    "esbuild-loader": "^4.2.2"
   },
   "browserslist": {
     "production": [
@@ -39,6 +44,6 @@
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
